添加用户信息页面添加

+++ performance-service/src/main/java/com/performance/service/UserLoginService.java
@@ -36,5 +36,5 @@ public interface UserLoginService {
      * @param loginId
      * @return
      */
-    Integer getCountByCreatedId(Long loginId);
+    Long getCountByCreatedId(Long loginId);
 }
